+++ packages/app/tests/application.spec.ts
@@ -6,6 +6,7 @@ import { BaseEvent, DataEvent, DataEventToken, EventDispatcher, eventDispatcher,
 import { cli, Command, Flag } from '../src/command.js';
 import { ClassType, isClass } from '@deepkit/core';
 import { ControllerConfig, ServiceContainer } from '../src/service-container.js';
+import { Stopwatch } from '@deepkit/stopwatch';
 
 Error.stackTraceLimit = 100;
 
@@ -460,7 +461,7 @@ test('service container hooks', () => {
         const m = new MyModule;
         const app = new ServiceContainer(new AppModule({ imports: [m] }));
         app.process();
-        expect(m.providersFound.length).toBe(8); //8 is the default, as the ServiceContainer adds default services
+        expect(m.providersFound.length).toBe(9); //9 is the default, as the ServiceContainer adds default services
         expect(m.controllersFound.length).toBe(0);
     }
 
@@ -471,7 +472,7 @@ test('service container hooks', () => {
         const m = new MyModule;
         const app = new ServiceContainer(new AppModule({ controllers: [Controller], imports: [m] }));
         app.process();
-        expect(m.providersFound.length).toBe(9);
+        expect(m.providersFound.length).toBe(10);
         expect(m.controllersFound.length).toBe(1);
     }
 
@@ -482,7 +483,7 @@ test('service container hooks', () => {
         const m = new MyModule;
         const app = new ServiceContainer(new AppModule({ providers: [Controller], imports: [m] }));
         app.process();
-        expect(m.providersFound.length).toBe(9);
+        expect(m.providersFound.length).toBe(10);
         expect(m.controllersFound.length).toBe(0);
     }
 
@@ -501,7 +502,7 @@ test('service container hooks', () => {
         const m = new MyModule;
         const app = new ServiceContainer(new AppModule({ imports: [baseModule, m] }));
         app.process();
-        expect(m.providersFound.length).toBe(10);
+        expect(m.providersFound.length).toBe(11);
         expect(m.controllersFound.length).toBe(1);
     }
 });