+++ packages/core/src/types.ts
@@ -1,2 +1,2 @@
-export type InjectMeta<T = never> = { __meta?: never & ['inject', T] }
-export type Inject<Type, Token = never> = Type & InjectMeta<Token>
+export type InjectMeta<T = never> = { __meta?: never & ['inject', T] };
+export type Inject<Type, Token = never> = Type & InjectMeta<Token>;