+++ packages/framework/src/module.ts
@@ -10,7 +10,6 @@
 
 import { ClassType, isClass, isPrototypeOfBase, ProcessLocker } from '@deepkit/core';
 import { EventDispatcher } from '@deepkit/event';
-import { mkdirSync } from 'fs';
 import { join } from 'path';
 import { ApplicationServer, ApplicationServerListener } from './application-server.js';
 import { DebugRouterController } from './cli/debug-router.js';
@@ -30,9 +29,9 @@ import { DatabaseListener } from './database/database-listener.js';
 import { Database, DatabaseRegistry } from '@deepkit/orm';
 import { MigrationCreateController, MigrationDownCommand, MigrationPendingCommand, MigrationProvider, MigrationUpCommand } from '@deepkit/sql/commands';
 import { FileStopwatchStore } from './debug/stopwatch/store.js';
-import { DebugDebugFramesCommand } from './cli/debug-debug-frames.js';
+import { DebugProfileFramesCommand } from './cli/debug-debug-frames.js';
 import { ConnectionWriter, rpcClass, RpcKernel, RpcKernelBaseConnection, RpcKernelConnection, RpcKernelSecurity, SessionState } from '@deepkit/rpc';
-import { AppConfigController } from './cli/app-config.js';
+import { DebugConfigController } from './cli/app-config.js';
 import { Zone } from './zone.js';
 import { DebugBrokerBus } from './debug/broker.js';
 import { ApiConsoleModule } from '@deepkit/api-console-module';
@@ -109,8 +108,8 @@ export class FrameworkModule extends createModule({
         ServerStartController,
         DebugRouterController,
         DebugDIController,
-        DebugDebugFramesCommand,
-        AppConfigController,
+        DebugProfileFramesCommand,
+        DebugConfigController,
 
         MigrationUpCommand,
         MigrationDownCommand,
@@ -136,6 +135,7 @@ export class FrameworkModule extends createModule({
         RpcKernelBaseConnection,
         ConnectionWriter,
 
+        BrokerDeepkitAdapter,
         BrokerCache,
         BrokerLock,
         BrokerQueue,
@@ -182,8 +182,6 @@ export class FrameworkModule extends createModule({
         }
 
         if (this.config.debug) {
-            mkdirSync(join(this.config.varPath, this.config.debugStorePath), { recursive: true });
-
             Zone.enable();
 
             this.addProvider({
@@ -203,32 +201,26 @@ export class FrameworkModule extends createModule({
 
             //only register the RPC controller
             this.addImport(new ApiConsoleModule({ listen: false, markdown: '' }).rename('internalApi'));
-
-            this.addListener(onAppShutdown.listen(async (
-                event, broker: DebugBrokerBus, store: StopwatchStore) => {
-                await store.close();
-                await broker.adapter.disconnect();
-            }));
-
             // this.setupProvider(LiveDatabase).enableChangeFeed(DebugRequest);
         }
 
+        this.addListener(onAppShutdown.listen(async (
+            event, broker: DebugBrokerBus, store: StopwatchStore) => {
+            await store.close();
+            await broker.adapter.disconnect();
+        }));
+
         this.addProvider(DebugBrokerBus);
-        this.addProvider(FileStopwatchStore);
-        this.addProvider({ provide: StopwatchStore, useExisting: FileStopwatchStore });
-        this.addProvider({
-            provide: Stopwatch,
-            useFactory(store: StopwatchStore, config: FrameworkConfig) {
-                const stopwatch = new Stopwatch(store);
-                if (config.profile || config.debug) {
-                    stopwatch.enable();
-                } else {
-                    stopwatch.disable();
-                }
-                return stopwatch;
-            }
-        });
-        this.addExport(Stopwatch);
+        this.addProvider({ provide: StopwatchStore, useClass: FileStopwatchStore });
+
+        const stopwatch = this.setupGlobalProvider<Stopwatch>();
+        if (this.config.profile || this.config.debug) {
+            stopwatch.enable();
+        } else {
+            stopwatch.disable();
+        }
+
+        this.addExport(Stopwatch, DebugBrokerBus, StopwatchStore);
     }
 
     postProcess() {