+++ packages/framework/tests/stopwatch.spec.ts
@@ -1,6 +1,8 @@
 import { expect, test } from '@jest/globals';
 import { decodeFrameData, deserializeFrameData, encodeFrameData } from '@deepkit/framework-debug-api';
-import { encodeCompoundKey, FrameCategory, FrameCategoryData } from '@deepkit/stopwatch';
+import { encodeCompoundKey, FrameCategory, FrameCategoryData, Stopwatch } from '@deepkit/stopwatch';
+import { App } from '@deepkit/app';
+import { FrameworkModule } from '../src/module.js';
 
 
 test('encode/decode', async () => {
@@ -17,4 +19,28 @@ test('encode/decode', async () => {
         console.log('http', http);
         expect(http).toEqual({ method: 'GET', clientIp: '127.0.0.1' });
     });
+});
+
+test('default disabled', () => {
+    const app = new App({
+        imports: [new FrameworkModule()]
+    });
+
+    expect(app.get(Stopwatch).active).toBe(false);
+});
+
+test('enabled with debug', () => {
+    const app = new App({
+        imports: [new FrameworkModule({ debug: true })]
+    });
+
+    expect(app.get(Stopwatch).active).toBe(true);
+});
+
+test('enabled with profile', () => {
+    const app = new App({
+        imports: [new FrameworkModule({ profile: true })]
+    });
+
+    expect(app.get(Stopwatch).active).toBe(true);
 });