+++ packages/app/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1-alpha.113](https://github.com/deepkit/deepkit-framework/compare/v1.0.1-alpha.112...v1.0.1-alpha.113) (2024-01-17)
+
+
+### Bug Fixes
+
+* **framework:** make sure profiler is only activated when debug/profile is true. ([c1d9357](https://github.com/deepkit/deepkit-framework/commit/c1d9357dba3b0a5568592fd5d61c8853bfe9b25e))
+
+
+
+
+
+## [1.0.1-alpha.112](https://github.com/deepkit/deepkit-framework/compare/v1.0.1-alpha.111...v1.0.1-alpha.112) (2024-01-16)
+
+
+### Features
+
+* **app:** improve CLI outputs/parsing by removing [@oclif](https://github.com/oclif) ([e38bbd1](https://github.com/deepkit/deepkit-framework/commit/e38bbd143daa2c856c57eca07a4fd29e884fe97e))
+
+
+
+
+
+## [1.0.1-alpha.111](https://github.com/deepkit/deepkit-framework/compare/v1.0.1-alpha.110...v1.0.1-alpha.111) (2024-01-15)
+
+**Note:** Version bump only for package @deepkit/app
+
+
+
+
+
 ## [1.0.1-alpha.110](https://github.com/deepkit/deepkit-framework/compare/v1.0.1-alpha.109...v1.0.1-alpha.110) (2024-01-11)
 
 **Note:** Version bump only for package @deepkit/app